Add add_sale_form.html
Add "Add Sale" button to "display_sales_by_customer.html".
Clicking on this button opens the "add_sale_form.html" page.
The customer object is loaded to the form on this page so the new
order will be added to this specific customer. After submitting
the new sale, the page is redirected to display sales by the
original customer.