Refactor service delete method to deleteById
Changed parameter form Customer to id. The controller delete method was
taking and id and retrieving a customer from the database and passing
that customer to the delete method in the service class. This change
stream lines all methods to use an integer of type long.

The delete method in the service now throws an CustomerServiceException if a customer
with specified id is not found.